requirements.txt
Flask
requests
psycopg2-binary
python-dotenv
PyJWT
